requirements.txt
# TO-DO: add python packages that are required for this collection
requests
PyGithub